Couple of quality of life changes
- Now on neotree keybindings it calls set relative line number so it has
automatic line numbering put on (the autocommands weren't working so
this was set to be done manually), this has patched this for all
commands on it. Not the nicest, but it works.
- qa action on the alpha buffer updated to only quit all when it is the
only open buffer. Basically ensures when someone opens the menu they
can't just quit all. on a qa.
- the q mapping on the alpha menu was also updated just to close the
buffer alpha is in instead of closing the entire session -> changed qa
to q.
- note if alpha is the last buffer and there are remaining modified
buffers, you will need to qa! or :Nuke from the menu
- still :Nuke will close the session from anywhere